Changed sidebar to autogenerated
To simplify the sidebar management, the configuration has been changed as follow:

Old config:
the structure was in one place alone in sidebar.js

New structure:
The sidebar is autogenerated and the positions, when needed, are managed in the files directly.
To generate a "group", every directory has a _category_.json file in it.