Refresh of the other/RKE2 installation.
I have added a demo of what runnign the command might look like. Using
vhs from https://github.com/charmbracelet/vhs. Liek ascinema but
different. The playback is specified in rke2.tape and then processed to
produce the output gif.

The gif is wrapped up in a details sections so its not being annoying
while your are trying to focus. An advantage is that it only starts
playing when the details is expanded. Helps avoid being lost in the
middle.

What do you think? More of this?

Signed-off-by: John Krug <john.krug@suse.com>